Add test for CKVSs
